requirements.txt
numpy
pandas
scikit-learn
ortools
statsmodels
xgboost
matplotlib
scipy
joblib
fastapi
uvicorn
python-dotenv
redis
psycopg2-binary
requests
